Change return type Interface Expression int to double & Implement SpecialVariables

+++ src/main/java/Grammar/Expression/IntLit.java
@@ -4,11 +4,10 @@ import GamePlay.Player;
 import GamePlay.Territory;
 
 import java.util.HashMap;
-import java.util.Map;
 
 public record IntLit(int val) implements Expression {
     @Override
-    public int eval(HashMap<String, Integer> bindings, Player p, Territory t) {
+    public double eval(HashMap<String, Integer> bindings, Player p, Territory t) {
         return val;
     }
     @Override
